Expand support for different share types
This is a breaking change. The current version of the code doesn't
handle public.image correctly in some cases. It assumes that the
value returned from getItemForTypeIdentifier is always a NSURL*.
This is true for things like Photos, but the app can define what
it passes back, while testing, I found that it can also be a
UIImage* instance or even NSData*, for example in Notes if you
create a note, add a drawing and insert a photo, the drawing will
be found as a public.image but it will be in NSData format, not
NSURL.

In the cases where we get UIImage* or NSData* in order to share them
later on we have to have a URL to them on disk, so I used app groups
to create a temporary container location to store the files. Then the
caller can reference these URLs in the JS code, and finally when we
close the share extension we clean up all the temp files to make sure
we don't keep growing the size of our disk footprint.